feat: Add controlled LoadZones component to shared dir
Move LoadZones subtree to src/components/TestOptions/LoadZones/ and
refactor to accept value/onChange props. Replace the original location
with a thin store bridge that reads from useGeneratorStore.
